+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/controller/ParamFlowRuleController.java
@@ -91,7 +91,9 @@ public class ParamFlowRuleController {
                                                                         @RequestParam String ip,
                                                                         @RequestParam Integer port) {
         AuthUser authUser = authService.getAuthUser(request);
-        authUser.authTarget(app, PrivilegeType.READ_RULE);
+        if (!authUser.authTarget(app, PrivilegeType.READ_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
         if (StringUtil.isEmpty(app)) {
             return Result.ofFail(-1, "app cannot be null or empty");
         }
@@ -130,7 +132,9 @@ public class ParamFlowRuleController {
     public Result<ParamFlowRuleEntity> apiAddParamFlowRule(HttpServletRequest request,
                                                            @RequestBody ParamFlowRuleEntity entity) {
         AuthUser authUser = authService.getAuthUser(request);
-        authUser.authTarget(entity.getApp(), PrivilegeType.WRITE_RULE);
+        if (!authUser.authTarget(entity.getApp(), PrivilegeType.WRITE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
         Result<ParamFlowRuleEntity> checkResult = checkEntityInternal(entity);
         if (checkResult != null) {
             return checkResult;
@@ -209,7 +213,9 @@ public class ParamFlowRuleController {
         if (oldEntity == null) {
             return Result.ofFail(-1, "id " + id + " does not exist");
         }
-        authUser.authTarget(oldEntity.getApp(), PrivilegeType.WRITE_RULE);
+        if (!authUser.authTarget(oldEntity.getApp(), PrivilegeType.WRITE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
         Result<ParamFlowRuleEntity> checkResult = checkEntityInternal(entity);
         if (checkResult != null) {
             return checkResult;
@@ -248,7 +254,9 @@ public class ParamFlowRuleController {
         if (oldEntity == null) {
             return Result.ofSuccess(null);
         }
-        authUser.authTarget(oldEntity.getApp(), PrivilegeType.DELETE_RULE);
+        if (!authUser.authTarget(oldEntity.getApp(), PrivilegeType.DELETE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
         try {
             repository.delete(id);
             publishRules(oldEntity.getApp(), oldEntity.getIp(), oldEntity.getPort()).get();
