+++ sentinel-dashboard/src/main/java/com/alibaba/csp/sentinel/dashboard/controller/v2/FlowControllerV2.java
@@ -73,7 +73,9 @@ public class FlowControllerV2 {
     @GetMapping("/rules")
     public Result<List<FlowRuleEntity>> apiQueryMachineRules(HttpServletRequest request, @RequestParam String app) {
         AuthUser authUser = authService.getAuthUser(request);
-        authUser.authTarget(app, PrivilegeType.READ_RULE);
+        if (!authUser.authTarget(app, PrivilegeType.READ_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
 
         if (StringUtil.isEmpty(app)) {
             return Result.ofFail(-1, "app can't be null or empty");
@@ -143,7 +145,9 @@ public class FlowControllerV2 {
     @PostMapping("/rule")
     public Result<FlowRuleEntity> apiAddFlowRule(HttpServletRequest request, @RequestBody FlowRuleEntity entity) {
         AuthUser authUser = authService.getAuthUser(request);
-        authUser.authTarget(entity.getApp(), PrivilegeType.WRITE_RULE);
+        if (!authUser.authTarget(entity.getApp(), PrivilegeType.WRITE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
 
         Result<FlowRuleEntity> checkResult = checkEntityInternal(entity);
         if (checkResult != null) {
@@ -180,7 +184,9 @@ public class FlowControllerV2 {
         if (entity == null) {
             return Result.ofFail(-1, "invalid body");
         }
-        authUser.authTarget(oldEntity.getApp(), PrivilegeType.WRITE_RULE);
+        if (!authUser.authTarget(oldEntity.getApp(), PrivilegeType.WRITE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
 
         entity.setApp(oldEntity.getApp());
         entity.setIp(oldEntity.getIp());
@@ -217,7 +223,9 @@ public class FlowControllerV2 {
         if (oldEntity == null) {
             return Result.ofSuccess(null);
         }
-        authUser.authTarget(oldEntity.getApp(), PrivilegeType.DELETE_RULE);
+        if (!authUser.authTarget(oldEntity.getApp(), PrivilegeType.DELETE_RULE)) {
+            return Result.ofFail(-1, "Can't operate because of no privilege");
+        }
         try {
             repository.delete(id);
             publishRules(oldEntity.getApp());
